Create CI Formatter - XLSX and CVS Reference
Version 2: CI Formatter - CSV and XLSX.py
Use this script if you are referencing a .csv you exported from the server against a .xlsx file. This is useful if you've already pulled an export from the server and want to narrow the list by referencing an xlsx. The .xlsx file should only contain the columns UserID and UserName.

Example:

A	B
1	jj76573	Jaiden Johnson
2	...	...
You will, again, be prompted to enter values for each field. These values will apply to ALL USERS in the list. This creates 13 columns and formats in UTF-8 .csv formatting.